+++ src/onegov.ballot/onegov/ballot/collections/__init__.py
@@ -1,0 +1,9 @@
+from onegov.ballot.collections.ballots import BallotCollection
+from onegov.ballot.collections.elections import ElectionCollection
+from onegov.ballot.collections.votes import VoteCollection
+
+__all__ = [
+    'BallotCollection',
+    'ElectionCollection',
+    'VoteCollection'
+]